demo: move counter demo services to pkg/service
